backend/requirements.txt
flask
flask-cors
tensorflow
transformers
torch
pytesseract
Pillow
numpy
scikit-learn
werkzeug
